requirements.txt
streamlit~=1.48.0
GitPython~=3.1.45
gitdb~=4.0.12
numpy~=2.3.2
pillow~=11.3.0
pip~=23.2.1
attrs~=25.3.0
wheel~=0.41.2
protobuf~=6.31.1
tornado~=6.5.1
Jinja2~=3.1.6
pytz~=2025.2
toml~=0.10.2
typing_extensions~=4.14.1
smmap~=5.0.2
mysql-connector-python~=9.4.0
charset-normalizer~=3.4.2
setuptools~=68.2.0
altair~=5.5.0
jsonschema~=4.25.0
narwhals~=2.0.1
pandas~=2.3.1
pyarrow~=21.0.0
packaging~=25.0
referencing~=0.36.2
click~=8.2.1
MarkupSafe~=3.0.2
python-dateutil~=2.9.0.post0
plotly~=6.2.0
pydeck~=0.9.1
certifi~=2025.8.3
tzdata~=2025.2
colorama~=0.4.6
urllib3~=2.5.0
six~=1.17.0
idna~=3.10
tenacity~=9.1.2
watchdog~=6.0.0
cachetools~=6.1.0
blinker~=1.9.0
websocket-client~=1.8.0
jsonschema-specifications~=2025.4.1
requests~=2.32.4
python-dotenv~=1.1.1